Flere fikser som følge av PR-tilbakemeldinger
